Frame.h: merge into Video.h.
The functions GetFrameBufferXXX() have been moved next to g_pFramebufferbits.
