+++ source/Debugger/Debugger_Display.cpp
@@ -35,7 +35,6 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "../Core.h"
 #include "../Interface.h"
 #include "../CPU.h"
-#include "../Frame.h"
 #include "../Windows/WinFrame.h"
 #include "../LanguageCard.h"
 #include "../Memory.h"
