--- source/Frame.cpp
@@ -1,66 +1,0 @@
-/*
-AppleWin : An Apple //e emulator for Windows
-
-Copyright (C) 1994-1996, Michael O'Brien
-Copyright (C) 1999-2001, Oliver Schmidt
-Copyright (C) 2002-2005, Tom Charlesworth
-Copyright (C) 2006-2014, Tom Charlesworth, Michael Pohoreski
-
-AppleWin is free software; you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation; either version 2 of the License, or
-(at your option) any later version.
-
-AppleWin is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with AppleWin; if not, write to the Free Software
-Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
-*/
-
-/* Description: Frame
- *
- * Author: Various
- */
-
-#include "StdAfx.h"
-
-#include "Frame.h"
-
-UINT GetFrameBufferBorderlessWidth(void)
-{
-	static const UINT uFrameBufferBorderlessW = 560;	// 560 = Double Hi-Res
-	return uFrameBufferBorderlessW;
-}
-
-UINT GetFrameBufferBorderlessHeight(void)
-{
-	static const UINT uFrameBufferBorderlessH = 384;	// 384 = Double Scan Line
-	return uFrameBufferBorderlessH;
-}
-
-// NB. These border areas are not visible (... and these border areas are unrelated to the 3D border below)
-UINT GetFrameBufferBorderWidth(void)
-{
-	static const UINT uBorderW = 20;
-	return uBorderW;
-}
-
-UINT GetFrameBufferBorderHeight(void)
-{
-	static const UINT uBorderH = 18;
-	return uBorderH;
-}
-
-UINT GetFrameBufferWidth(void)
-{
-	return GetFrameBufferBorderlessWidth() + 2*GetFrameBufferBorderWidth();
-}
-
-UINT GetFrameBufferHeight(void)
-{
-	return GetFrameBufferBorderlessHeight() + 2*GetFrameBufferBorderHeight();
-}
