--- source/Frame.h
@@ -1,8 +1,0 @@
-#pragma once
-
-	UINT GetFrameBufferBorderlessWidth(void);
-	UINT GetFrameBufferBorderlessHeight(void);
-	UINT GetFrameBufferBorderWidth(void);
-	UINT GetFrameBufferBorderHeight(void);
-	UINT GetFrameBufferWidth(void);
-	UINT GetFrameBufferHeight(void);
