+++ source/RGBMonitor.cpp
@@ -3,7 +3,6 @@
 #include "StdAfx.h"
 
 #include "RGBMonitor.h"
-#include "Frame.h"
 #include "Memory.h" // MemGetMainPtr() MemGetAuxPtr()
 #include "Video.h"
 #include "Card.h"
