+++ source/Video.cpp
@@ -31,7 +31,6 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "Video.h"
 #include "Core.h"
 #include "CPU.h"
-#include "Frame.h"
 #include "Interface.h"
 #include "Log.h"
 #include "Memory.h"
@@ -128,7 +127,43 @@ static bool g_bVideoScannerNTSC = true;  // NTSC video scanning (or PAL)
 // ----- ALL GLOBALLY ACCESSIBLE FUNCTIONS ARE BELOW THIS LINE -----
 //
 
+UINT GetFrameBufferBorderlessWidth(void)
+{
+	static const UINT uFrameBufferBorderlessW = 560;	// 560 = Double Hi-Res
+	return uFrameBufferBorderlessW;
+}
+
+UINT GetFrameBufferBorderlessHeight(void)
+{
+	static const UINT uFrameBufferBorderlessH = 384;	// 384 = Double Scan Line
+	return uFrameBufferBorderlessH;
+}
+
+// NB. These border areas are not visible (... and these border areas are unrelated to the 3D border below)
+UINT GetFrameBufferBorderWidth(void)
+{
+	static const UINT uBorderW = 20;
+	return uBorderW;
+}
+
+UINT GetFrameBufferBorderHeight(void)
+{
+	static const UINT uBorderH = 18;
+	return uBorderH;
+}
+
+UINT GetFrameBufferWidth(void)
+{
+	return GetFrameBufferBorderlessWidth() + 2 * GetFrameBufferBorderWidth();
+}
+
+UINT GetFrameBufferHeight(void)
+{
+	return GetFrameBufferBorderlessHeight() + 2 * GetFrameBufferBorderHeight();
+}
+
 //===========================================================================
+
 void VideoReinitialize (bool bInitVideoScannerAddress /*= true*/)
 {
 	NTSC_VideoReinitialize( g_dwCyclesThisFrame, bInitVideoScannerAddress );
